Adding namespace to all resource classes.
