+++ lib/CardList.php
@@ -1,0 +1,50 @@
+<?php
+namespace Omise;
+
+use Omise\Res\OmiseApiResource;
+use Omise\Card;
+
+class CardList extends OmiseApiResource
+{
+    const ENDPOINT = 'cards';
+
+    private $_customerID;
+  
+    /**
+     * @param array  $cards
+     * @param string $customerID
+     * @param string $publickey
+     * @param string $secretkey
+     */
+    public function __construct($cards, $customerID, $publickey = null, $secretkey = null)
+    {
+        parent::__construct($publickey, $secretkey);
+        $this->_customerID = $customerID;
+        $this->refresh($cards);
+    }
+  
+    /**
+     * retrieve a card
+     *
+     * @param  string $id
+     *
+     * @return OmiseCard
+     */
+    public function retrieve($id)
+    {
+        $result = $this->apiRequestor->get($this->getUrl($id), self::getResourceKey());
+
+        return new Card($result, $this->_customerID, $this->_publickey, $this->_secretkey);
+    }
+  
+
+    /**
+     * @param  string $id
+     *
+     * @return string
+     */
+    private function getUrl($id = '')
+    {
+        return \Omise\ApiRequestor::OMISE_API_URL . 'customers/' . $this->_customerID . '/' . self::ENDPOINT . '/' . $id;
+    }
+}
