+++ lib/ScheduleList.php
@@ -1,0 +1,18 @@
+<?php
+namespace Omise;
+
+use Omise\Res\OmiseApiResource;
+use Omise\Schedule;
+
+class ScheduleList extends OmiseApiResource
+{
+    /**
+     * @param  string $id
+     *
+     * @return OmiseOccurrence
+     */
+    public function retrieve($id)
+    {
+        return Schedule::retrieve($id, $this->_publickey, $this->_secretkey);
+    }
+}
