+++ lib/omise/OmiseAccount.php
@@ -1,8 +1,6 @@
 <?php
 /**
  * @deprecated 3.0.0  No longer used this class and will be removed on Jan 4th, 2020.
+ * phpcs:disable
  */
-class OmiseAccount extends \Omise\Account
-{
-
-}
+class OmiseAccount extends \Omise\Account { }
