+++ lib/omise/OmiseCard.php
@@ -1,75 +1,6 @@
 <?php
-use Omise\Res\OmiseApiResource;
-
-class OmiseCard extends OmiseApiResource
-{
-    const ENDPOINT = 'cards';
-
-    private $_customerID;
-
-    /**
-     * Object representing a card. Cards are retrieved using a `Customer`.
-     *
-     * @param array  $array
-     * @param string $customerID
-     * @param string $publickey
-     * @param string $secretkey
-     */
-    public function __construct($array, $customerID, $publickey = null, $secretkey = null)
-    {
-        parent::__construct($publickey, $secretkey);
-
-        $this->_customerID = $customerID;
-        $this->refresh($array);
-    }
-
-    /**
-     * (non-PHPdoc)
-     *
-     * @see OmiseApiResource::g_reload()
-     */
-    public function reload()
-    {
-        parent::g_reload($this->getUrl($this['id']));
-    }
-
-    /**
-     * (non-PHPdoc)
-     *
-     * @see OmiseApiResource::g_update()
-     */
-    public function update($params)
-    {
-        parent::g_update($this->getUrl($this['id']), $params);
-    }
-
-    /**
-     * (non-PHPdoc)
-     *
-     * @see OmiseApiResource::g_destroy()
-     */
-    public function destroy()
-    {
-        parent::g_destroy($this->getUrl($this['id']));
-    }
-
-    /**
-     * (non-PHPdoc)
-     *
-     * @see OmiseApiResource::isDestroyed()
-     */
-    public function isDestroyed()
-    {
-        return parent::isDestroyed();
-    }
-
-    /**
-     * @param  string $cardID
-     *
-     * @return string
-     */
-    private function getUrl($cardID = '')
-    {
-        return \Omise\ApiRequestor::OMISE_API_URL.OmiseCustomer::ENDPOINT.'/'.$this->_customerID.'/'.self::ENDPOINT.'/'.$cardID;
-    }
-}
+/**
+ * @deprecated 3.0.0  No longer used this class and will be removed on Jan 4th, 2020.
+ * phpcs:disable
+ */
+class OmiseCard extends \Omise\Card { }
