+++ lib/omise/OmiseCardList.php
@@ -1,47 +1,6 @@
 <?php
-use Omise\Res\OmiseApiResource;
-
-class OmiseCardList extends OmiseApiResource
-{
-    const ENDPOINT = 'cards';
-
-    private $_customerID;
-  
-    /**
-     * @param array  $cards
-     * @param string $customerID
-     * @param string $publickey
-     * @param string $secretkey
-     */
-    public function __construct($cards, $customerID, $publickey = null, $secretkey = null)
-    {
-        parent::__construct($publickey, $secretkey);
-        $this->_customerID = $customerID;
-        $this->refresh($cards);
-    }
-  
-    /**
-     * retrieve a card
-     *
-     * @param  string $id
-     *
-     * @return OmiseCard
-     */
-    public function retrieve($id)
-    {
-        $result = $this->apiRequestor->get($this->getUrl($id), self::getResourceKey());
-
-        return new OmiseCard($result, $this->_customerID, $this->_publickey, $this->_secretkey);
-    }
-  
-
-    /**
-     * @param  string $id
-     *
-     * @return string
-     */
-    private function getUrl($id = '')
-    {
-        return \Omise\ApiRequestor::OMISE_API_URL.'customers/'.$this->_customerID.'/'.self::ENDPOINT.'/'.$id;
-    }
-}
+/**
+ * @deprecated 3.0.0  No longer used this class and will be removed on Jan 4th, 2020.
+ * phpcs:disable
+ */
+class OmiseCardList extends \Omise\CardList { }
