+++ lib/omise/OmiseCustomer.php
@@ -1,151 +1,6 @@
 <?php
-use Omise\Res\OmiseApiResource;
-
-class OmiseCustomer extends OmiseApiResource
-{
-    const ENDPOINT = 'customers';
-
-    /**
-     * Retrieves a customer.
-     *
-     * @param  string $id
-     * @param  string $publickey
-     * @param  string $secretkey
-     *
-     * @return OmiseCustomer
-     */
-    public static function retrieve($id = '', $publickey = null, $secretkey = null)
-    {
-        return parent::g_retrieve(get_class(), self::getUrl($id), $publickey, $secretkey);
-    }
-
-    /**
-     * Search for customers.
-     *
-     * @param  string $query
-     * @param  string $publickey
-     * @param  string $secretkey
-     *
-     * @return OmiseSearch
-     */
-    public static function search($query = '', $publickey = null, $secretkey = null)
-    {
-        return OmiseSearch::scope('customer', $publickey, $secretkey)->query($query);
-    }
-
-    /**
-     * Creates a new customer.
-     *
-     * @param  array  $params
-     * @param  string $publickey
-     * @param  string $secretkey
-     *
-     * @return OmiseCustomer
-     */
-    public static function create($params, $publickey = null, $secretkey = null)
-    {
-        return parent::g_create(get_class(), self::getUrl(), $params, $publickey, $secretkey);
-    }
-
-    /**
-     * (non-PHPdoc)
-     *
-     * @see OmiseApiResource::g_reload()
-     */
-    public function reload()
-    {
-        if ($this['object'] === 'customer') {
-            parent::g_reload(self::getUrl($this['id']));
-        } else {
-            parent::g_reload(self::getUrl());
-        }
-    }
-
-    /**
-     * (non-PHPdoc)
-     *
-     * @see OmiseApiResource::g_update()
-     */
-    public function update($params)
-    {
-        parent::g_update(self::getUrl($this['id']), $params);
-    }
-
-    /**
-     * (non-PHPdoc)
-     *
-     * @see OmiseApiResource::g_destroy()
-     */
-    public function destroy()
-    {
-        parent::g_destroy(self::getUrl($this['id']));
-    }
-
-    /**
-     * (non-PHPdoc)
-     *
-     * @see OmiseApiResource::isDestroyed()
-     */
-    public function isDestroyed()
-    {
-        return parent::isDestroyed();
-    }
-
-    /**
-     * Gets a list of all cards belongs to this customer.
-     *
-     * @param  array $options
-     *
-     * @return OmiseCardList
-     */
-    public function cards($options = array())
-    {
-        if (is_array($options) && ! empty($options)) {
-            $cards = $this->apiRequestor->get(self::getUrl($this['id']) . '/cards?' . http_build_query($options), parent::getResourceKey());
-        } else {
-            $cards = $this['cards'];
-        }
-
-        return new OmiseCardList($cards, $this['id'], $this->_publickey, $this->_secretkey);
-    }
-  
-    /**
-     * cards() alias
-     *
-     * @deprecated deprecated since version 2.0.0 use '$customer->cards()'
-     *
-     * @return     OmiseCardList
-     */
-    public function getCards($options = array())
-    {
-        return $this->cards($options);
-    }
-
-    /**
-     * Gets a list of charge schedules that belongs to a given customer.
-     *
-     * @param  array|string $options
-     *
-     * @return OmiseScheduleList
-     */
-    public function schedules($options = array())
-    {
-        if ($this['object'] === 'customer') {
-            if (is_array($options)) {
-                $options = '?' . http_build_query($options);
-            }
-
-            return parent::g_retrieve('OmiseScheduleList', self::getUrl($this['id'] . '/schedules' . $options), $this->_publickey, $this->_secretkey);
-        }
-    }
-
-    /**
-     * @param  string $id
-     *
-     * @return string
-     */
-    private static function getUrl($id = '')
-    {
-        return \Omise\ApiRequestor::OMISE_API_URL.self::ENDPOINT.'/'.$id;
-    }
-}
+/**
+ * @deprecated 3.0.0  No longer used this class and will be removed on Jan 4th, 2020.
+ * phpcs:disable
+ */
+class OmiseCustomer extends \Omise\Customer { }
