+++ lib/omise/OmiseOccurrence.php
@@ -1,36 +1,6 @@
 <?php
-use Omise\Res\OmiseApiResource;
-
-class OmiseOccurrence extends OmiseApiResource
-{
-    const ENDPOINT = 'occurrences';
-
-    /**
-     * Retrieves an occurence object.
-     *
-     * @param  string $id
-     * @param  string $publickey
-     * @param  string $secretkey
-     *
-     * @return OmiseOccurrence
-     */
-    public static function retrieve($id, $publickey = null, $secretkey = null)
-    {
-        return parent::g_retrieve(get_class(), self::getUrl($id), $publickey, $secretkey);
-    }
-
-    public function reload()
-    {
-        parent::g_reload(self::getUrl($this['id']));
-    }
-
-    /**
-     * @param  string $id
-     *
-     * @return string
-     */
-    private static function getUrl($id = '')
-    {
-        return \Omise\ApiRequestor::OMISE_API_URL.self::ENDPOINT . '/' . $id;
-    }
-}
+/**
+ * @deprecated 3.0.0  No longer used this class and will be removed on Jan 4th, 2020.
+ * phpcs:disable
+ */
+class OmiseOccurrence extends \Omise\Occurrence { }
