+++ lib/omise/OmiseOccurrenceList.php
@@ -1,15 +1,6 @@
 <?php
-use Omise\Res\OmiseApiResource;
-
-class OmiseOccurrenceList extends OmiseApiResource
-{
-    /**
-     * @param  string $id
-     *
-     * @return OmiseOccurrence
-     */
-    public function retrieve($id)
-    {
-        return OmiseOccurrence::retrieve($id, $this->_publickey, $this->_secretkey);
-    }
-}
+/**
+ * @deprecated 3.0.0  No longer used this class and will be removed on Jan 4th, 2020.
+ * phpcs:disable
+ */
+class OmiseOccurrenceList extends \Omise\OccurrenceList { }
