+++ lib/omise/OmiseSource.php
@@ -1,29 +1,6 @@
 <?php
-use Omise\Res\OmiseApiResource;
-
-class OmiseSource extends OmiseApiResource
-{
-    const ENDPOINT = 'sources';
-
-    /**
-     * Creates a new source.
-     *
-     * @param  array  $params
-     * @param  string $publickey
-     * @param  string $secretkey
-     *
-     * @return OmiseSource
-     */
-    public static function create($params, $publickey = null, $secretkey = null)
-    {
-        return parent::g_create(get_class(), self::getUrl(), $params, $publickey, $secretkey);
-    }
-
-    /**
-     * @return string
-     */
-    private static function getUrl()
-    {
-        return \Omise\ApiRequestor::OMISE_API_URL.self::ENDPOINT;
-    }
-}
+/**
+ * @deprecated 3.0.0  No longer used this class and will be removed on Jan 4th, 2020.
+ * phpcs:disable
+ */
+class OmiseSource extends \Omise\Source { }
